requirements.txt
randomname==0.2.1
httpx==0.27.0
pytimeparse==1.1.8
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"